backend/auth-service/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.5.0
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.10
pyotp==2.9.0
qrcode==7.4.2
argon2-cffi==23.1.0
python-dotenv==1.0.1
Pillow==10.4.0
drf-yasg==1.21.7
